Add files via upload
Vinamilk is currently the largest dairy company in Vietnam, with the ticker 'VNM' on HOSE (Ho Chi Minh City Stock Exchange).
This is a part of my project about Vinamilk's stock analysis at university. It is a prediction of Vinamilk's stock open and close prices for the next 6 months. I use 5 years of historical data from 2/1/2019 to 22/11/2024, in which I train my model - BiLSTM on 80% of the dataset and test on the remaining 20%. The model will look at the patterns of stock prices over a sliding window of 50 time steps and predict the next price. After training BiLSTM on my data, the accuracy result on the test period is good. Thus, I decide to use the model to predict further 6 months (25/11/2024 - 01/08/2024), and the prediction indicates an increase of the price from 64k to 66.6k. This indicates a positive growth of Vinamilk's stock, and thus investors can base on this to be optimistic about VNM.