Add files via upload
This is a project of Employee Performance Analysis, in which I was the leader and responsible for all the coding, planning outline, and tracking the content of my teamates. After finishing the code file, we analysed the correlation & got insights from charts in the EDA, then used ML models (SVM, RF, BRF, and MLPClassifier) to classify the performance rating, and finally pinpointed top 3 drivers of our target (performance). We also conducted some recommendations to improve employee performance (all in the "Employee Performance Analysis_Report" file). Please have a look on it!
*Note: Due to the time constraints when coding at the time, I did not plot feature importances after running the RF models. Thus, if I had had the opportunity to run the project again, I would have absolutely plotted such a chart to further ensure our top 3 performance drivers were correctly chosen.